initial implementation of the NotificationManager
